Minor refactoring
This PR is mainly regarding the code design as I have little experience with JWT. It is generally better to use `is.*(x)` or `inherits(x, *)` than `class(x) == *` as it always returns a single boolean. Further I have made changes to use the logging infrastructure in `fiery`. Lastly I've refactored the `check_jwt()` method a bit to always return a boolean.